+++ pallets/staking/src/lib.rs
@@ -111,43 +111,40 @@ pub mod pallet {
     #[pallet::without_storage_info]
     pub struct Pallet<T>(_);
 
-    /// Stores the relationship between
-    /// a validator's stash account and their threshold server's sr25519 and x25519 keys.
+    /// Stores the relationship between a validator's stash account and the information about their
+    /// threshold server.
     ///
-    /// Clients query this via state or `stakingExtension_getKeys` RPC and uses
-    /// the x25519 pub key in noninteractive ECDH for authenticating/encrypting distribute TSS
-    /// shares over HTTP.
+    /// # Note
+    ///
+    /// This mapping doesn't only include information about validators in the active set, but also
+    /// information about validator candidates (i.e, those _might_ be part of the active set in the
+    /// following era).
     #[pallet::storage]
     #[pallet::getter(fn threshold_server)]
-    pub type ThresholdServers<T: Config> = StorageMap<
-        _,
-        Blake2_128Concat,
-        <T as pallet_session::Config>::ValidatorId,
-        ServerInfo<T::AccountId>,
-        OptionQuery,
-    >;
+    pub type ThresholdServers<T: Config> =
+        StorageMap<_, Blake2_128Concat, T::ValidatorId, ServerInfo<T::AccountId>, OptionQuery>;
 
+    /// A mapping between a threshold server's Account ID and its corresponding validator's stash
+    /// account (i.e the reverse of [ThresholdServers]).
+    ///
+    /// # Note
+    ///
+    /// This mapping doesn't only include information about validators in the active set, but also
+    /// information about validator candidates (i.e, those _might_ be part of the active set in the
+    /// following era).
     #[pallet::storage]
     #[pallet::getter(fn threshold_to_stash)]
-    pub type ThresholdToStash<T: Config> = StorageMap<
-        _,
-        Blake2_128Concat,
-        T::AccountId,
-        <T as pallet_session::Config>::ValidatorId,
-        OptionQuery,
-    >;
+    pub type ThresholdToStash<T: Config> =
+        StorageMap<_, Blake2_128Concat, T::AccountId, T::ValidatorId, OptionQuery>;
 
-    /// Stores the relationship between a signing group (u8) and its member's (validator's)
-    /// threshold server's account.
+    /// Keeps track of all the validators in a particular subgroup.
+    ///
+    /// Only active validators (so not candiates) should be assigned a subgroup and be included in
+    /// this mapping.
     #[pallet::storage]
     #[pallet::getter(fn signing_groups)]
-    pub type SigningGroups<T: Config> = StorageMap<
-        _,
-        Blake2_128Concat,
-        SubgroupId,
-        Vec<<T as pallet_session::Config>::ValidatorId>,
-        OptionQuery,
-    >;
+    pub type SigningGroups<T: Config> =
+        StorageMap<_, Blake2_128Concat, SubgroupId, Vec<T::ValidatorId>, OptionQuery>;
 
     /// Mapping between a validator and their assigned subgroup for the given session.
     ///
@@ -350,45 +347,45 @@ pub mod pallet {
             Ok(().into())
         }
 
-        /// Wraps's substrate validate but forces threshold key and endpoint
-        /// `endpoint`: nodes's endpoint
-        /// `threshold_account`: nodes's threshold account
+        /// Wrap's Substrate's `staking_pallet::validate()` extrinsic, but enforces that
+        /// information about a validator's threshold server is provided.
+        ///
+        /// Note that - just like the original `validate()` extrinsic - the effects of this are
+        /// only applied in the following era.
         #[pallet::call_index(3)]
         #[pallet::weight(<T as Config>::WeightInfo::validate())]
         pub fn validate(
             origin: OriginFor<T>,
             prefs: ValidatorPrefs,
-            endpoint: Vec<u8>,
-            tss_account: T::AccountId,
-            x25519_public_key: X25519PublicKey,
+            server_info: ServerInfo<T::AccountId>,
         ) -> DispatchResult {
             let who = ensure_signed(origin.clone())?;
+
             ensure!(
-                endpoint.len() as u32 <= T::MaxEndpointLength::get(),
+                server_info.endpoint.len() as u32 <= T::MaxEndpointLength::get(),
                 Error::<T>::EndpointTooLong
             );
 
             ensure!(
-                !ThresholdToStash::<T>::contains_key(&tss_account),
+                !ThresholdToStash::<T>::contains_key(&server_info.tss_account),
                 Error::<T>::TssAccountAlreadyExists
             );
 
-            let stash = Self::get_stash(&who)?;
             pallet_staking::Pallet::<T>::validate(origin, prefs)?;
-            let validator_id = <T as pallet_session::Config>::ValidatorId::try_from(stash)
-                .or(Err(Error::<T>::InvalidValidatorId))?;
 
-            ThresholdServers::<T>::insert(
-                &validator_id,
-                ServerInfo {
-                    tss_account: tss_account.clone(),
-                    x25519_public_key,
-                    endpoint: endpoint.clone(),
-                },
-            );
-            ThresholdToStash::<T>::insert(&tss_account, validator_id);
+            let stash = Self::get_stash(&who)?;
+            let validator_id =
+                T::ValidatorId::try_from(stash).or(Err(Error::<T>::InvalidValidatorId))?;
 
-            Self::deposit_event(Event::NodeInfoChanged(who, endpoint, tss_account));
+            ThresholdServers::<T>::insert(&validator_id, server_info.clone());
+            ThresholdToStash::<T>::insert(&server_info.tss_account, validator_id);
+
+            Self::deposit_event(Event::NodeInfoChanged(
+                who,
+                server_info.endpoint,
+                server_info.tss_account,
+            ));
+
             Ok(())
         }