+++ pallets/staking/src/tests.rs
@@ -48,36 +48,43 @@ fn it_takes_in_an_endpoint() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info,
         ));
+
         let ServerInfo { tss_account, endpoint, .. } = Staking::threshold_server(1).unwrap();
         assert_eq!(endpoint, vec![20]);
         assert_eq!(tss_account, 3);
         assert_eq!(Staking::threshold_to_stash(3).unwrap(), 1);
+
+        let server_info = ServerInfo {
+            tss_account: 3,
+            x25519_public_key: NULL_ARR,
+            endpoint: vec![20, 20, 20, 20],
+        };
         assert_noop!(
             Staking::validate(
                 RuntimeOrigin::signed(4),
                 pallet_staking::ValidatorPrefs::default(),
-                vec![20, 20, 20, 20],
-                3,
-                NULL_ARR
+                server_info,
             ),
             Error::<Test>::EndpointTooLong
         );
+
+        let server_info =
+            ServerInfo { tss_account: 5, x25519_public_key: NULL_ARR, endpoint: vec![20, 20] };
         assert_noop!(
             Staking::validate(
                 RuntimeOrigin::signed(4),
                 pallet_staking::ValidatorPrefs::default(),
-                vec![20, 20],
-                5,
-                NULL_ARR
+                server_info
             ),
-            Error::<Test>::NotController
+            pallet_staking::Error::<Test>::NotController
         );
     });
 }
@@ -90,12 +97,13 @@ fn it_will_not_allow_validator_to_use_existing_tss_account() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info.clone(),
         ));
 
         // Attempt to call validate with a TSS account which already exists
@@ -108,9 +116,7 @@ fn it_will_not_allow_validator_to_use_existing_tss_account() {
             Staking::validate(
                 RuntimeOrigin::signed(2),
                 pallet_staking::ValidatorPrefs::default(),
-                vec![20],
-                3,
-                NULL_ARR
+                server_info,
             ),
             Error::<Test>::TssAccountAlreadyExists
         );
@@ -125,12 +131,13 @@ fn it_changes_endpoint() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info,
         ));
 
         assert_ok!(Staking::change_endpoint(RuntimeOrigin::signed(1), vec![30]));
@@ -151,12 +158,13 @@ fn it_changes_threshold_account() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info,
         ));
 
         assert_ok!(Staking::change_threshold_accounts(RuntimeOrigin::signed(1), 4, NULL_ARR));
@@ -174,12 +182,13 @@ fn it_changes_threshold_account() {
             100u64,
             pallet_staking::RewardDestination::Account(2),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 5, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            5,
-            NULL_ARR
+            server_info,
         ));
 
         assert_noop!(
@@ -197,12 +206,13 @@ fn it_will_not_allow_existing_tss_account_when_changing_threshold_account() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info,
         ));
 
         // Check that we cannot change to a TSS account which already exists
@@ -211,12 +221,13 @@ fn it_will_not_allow_existing_tss_account_when_changing_threshold_account() {
             100u64,
             pallet_staking::RewardDestination::Account(2),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 5, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            5,
-            NULL_ARR
+            server_info,
         ));
 
         assert_noop!(
@@ -235,12 +246,13 @@ fn it_deletes_when_no_bond_left() {
             100u64,
             pallet_staking::RewardDestination::Account(1),
         ));
+
+        let server_info =
+            ServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
-            vec![20],
-            3,
-            NULL_ARR
+            server_info,
         ));
 
         let ServerInfo { tss_account, endpoint, .. } = Staking::threshold_server(2).unwrap();