+++ CHANGELOG.md
@@ -23,6 +23,11 @@ At the moment this project **does not** adhere to
   `ProgramInfo` was renamed to `interface_description` to be more semantically accurate. This field
   will now be used to describe program interfaces, including the auxilary and configuration
   interfaces of the program.
+- In [#658](https://github.com/entropyxyz/entropy-core/pull/658), `RegisteredInfo` and
+  `RegisteringDetails` now contain `version_number`. As well `KeyVersionNumber` was added as a
+  config parameter to the `Relayer` pallet.
+- In [#659](https://github.com/entropyxyz/entropy-core/pull/659), the Staking Extension pallet's
+  `validate` extrinsic changed to take a `ServerInfo` struct instead of individual fields.
 
 ### Added
 - Add ValidatorSubgroupRotated event ([#618](https://github.com/entropyxyz/entropy-core/pull/618))
@@ -34,6 +39,7 @@ At the moment this project **does not** adhere to
 - Add proactive refresh keys on-chain ([#629](https://github.com/entropyxyz/entropy-core/pull/629))
 - Change test-cli default access mode and update readme for recent changes ([#643](https://github.com/entropyxyz/entropy-core/pull/643))
 - Disallow using existing TSS account IDs in Staking pallet ([#657](https://github.com/entropyxyz/entropy-core/pull/657))
+- Clean ups around Staking Extension's `validate()` extrinsic ([#659](https://github.com/entropyxyz/entropy-core/pull/659))
 
 ## [0.0.10](https://github.com/entropyxyz/entropy-core/compare/release/v0.0.9...release/v0.0.10) - 2024-01-24